+++ tests/test_draft.py
@@ -13,34 +13,8 @@ class Custom:
         pass
 
 
-def test_draft_from_config_with_auto_draft(registry):
-    @registry.misc_bis.register("partial", auto_draft_in_config=True)
-    def partial_function(required_val: int, param: int = 3) -> Custom:
-        return Custom(required_val + param)
-
-    config = """
-[model]
-@misc_bis = "partial"
-param = 3
-"""
-    obj = Config().from_str(config).resolve(registry=registry)["model"]
-    assert isinstance(obj, Draft)
-    with pytest.raises(TypeError):
-        obj.foo()
-    with pytest.raises(TypeError):
-        obj()
-    with pytest.raises(TypeError):
-        obj + 5
-    assert Draft.instantiate(obj, required_val=2).value == 5
-
-    obj2 = partial_function(param=5, required_val=2)
-    assert obj2.value == 7
-
-    assert Draft.instantiate(obj2, required_val=2).value == 7
-
-
-def test_draft_from_config_explicit(registry):
-    @registry.misc_bis.register("partial_explicit")
+def test_draft_from_config(registry):
+    @registry.misc_bis.register("partial")
     def partial_function(required_val: int, param: int = 3) -> Custom:
         return Custom(required_val + param)
 
@@ -69,15 +43,13 @@ def test_draft_from_python(registry):
     def partial_function(required_val: int, param: int = 3) -> Custom:
         return Custom(required_val + param)
 
-    partial_function = registry.misc_bis.register(
-        "partial", func=partial_function, auto_draft_in_config=True
-    )
+    partial_function = registry.misc_bis.register("partial", func=partial_function)
 
     @validate_arguments
     def use_draft(draft: Draft[Custom]) -> Custom:
         return draft.instantiate(required_val=2)
 
-    assert partial_function.draft(required_val=2).value == 5
+    assert partial_function.draft(required_val=2).instantiate().value == 5
 
     obj = partial_function.draft(param=3)
     assert use_draft(obj).value == 5
@@ -102,7 +74,7 @@ def test_draft_from_type(registry):
         def __init__(self, value: int):
             self.value = value
 
-    MyClass = registry.misc_bis.register("partial", auto_draft_in_config=True)(MyClass)
+    MyClass = registry.misc_bis.register("partial")(MyClass)
     assert isinstance(MyClass.draft(value=3), Draft)
 
     obj = MyClass.draft()
@@ -117,9 +89,7 @@ def test_draft_error_from_function(registry):
     def partial_function(required_val: int, param: int = 3) -> CustomError:
         return CustomError(required_val + param)
 
-    partial_function = registry.misc_bis.register("partial", auto_draft_in_config=True)(
-        partial_function
-    )
+    partial_function = registry.misc_bis.register("partial")(partial_function)
 
     obj = partial_function.draft()
 
@@ -144,9 +114,7 @@ def test_draft_error_from_type(registry):
         def __init__(self, value: int):
             self.value = value
 
-    CustomError = registry.misc_bis.register("partial", auto_draft_in_config=True)(
-        CustomError
-    )
+    CustomError = registry.misc_bis.register("partial")(CustomError)
 
     obj = CustomError.draft()