+++ changelog.md
@@ -2,12 +2,13 @@
 
 ## Unreleased
 
-- Deprecate `auto_draft_in_config`, in favor of explicitly using
+- Drop `auto_draft_in_config`, in favor of explicitly using
   ```
   @registry: myname !draft
   ```
   to define a draft object
-- :boom: Even if a `MyClass.draft(...)` instantiation contains all required parameter, it will still return a draft. `draft.instantiate` must be called to instantiate the actual object.
+- :boom: Even if a `MyClass.draft(...)` instantiation contains all required parameter, it will now still return a draft. `draft.instantiate` must be called to instantiate the actual object.
+- Support `isinstance(obj, Draft)` for Draft instances
 
 ## v0.9.0 (2025-05-15)